orders/src/routes/new.ts
import express, { Request, Response } from 'express';
import {
  OrderStatus,
  requireAuth,
  validateRequest,
  BadRequestError,
} from '@weitickets/common';
import { body } from 'express-validator';
import mongoose from 'mongoose';
import { Order } from '../models/order';
import { Commodity } from '../models/commodity';
import { NotFoundError } from '@weitickets/common';
import { OrderCreatedPublisher } from '../events/publishers/order-created-publisher';
import { natsWrapper } from '../nats-wrapper';

const router = express.Router();

const EXPIRATION_WINDOW_SECONDS = 1 * 60;

router.post(
  '/api/orders',
  requireAuth,
  [
    body('ticketId')
      .notEmpty()
      .custom((input: string) => mongoose.Types.ObjectId.isValid(input))
      .withMessage('Ticket id is required'),
  ],
  validateRequest,
  async (req: Request, res: Response) => {
    const { ticketId } = req.body;

    const commodity = await Commodity.findById(ticketId);
    if (!commodity) {
      throw new NotFoundError();
    }

    const isReserved = await commodity.isReserved(); // 定義在mongoose裡面
    if (isReserved) {
      throw new BadRequestError('Ticket is already reserved');
    }

    const expiration = new Date();
    expiration.setSeconds(expiration.getSeconds() + EXPIRATION_WINDOW_SECONDS);

    const order = Order.build({
      userId: req.currentUser!.id,
      status: OrderStatus.Created,
      expiresAt: expiration,
      ticket: commodity,
    });
    await order.save();

    new OrderCreatedPublisher(natsWrapper.client).publish({
      id: order.id,
      version: order.version,
      userId: order.userId,
      status: order.status,
      expiresAt: order.expiresAt.toISOString(),
      ticket: {
        id: commodity.id,
        price: commodity.price,
      },
    })

    res.status(201).send(order);
  }
);

export { router as newOrdersRouter };
